+++ src/main.cpp
@@ -10,8 +10,8 @@ class ExampleSolver : public Solver {
 
 // Sample Implemeneation
 void ExampleSolver::solve(Board &board) {
-    for (int i = 0; i < board.getX(); i++) {
-        for (int j = 0; j < board.getY(); j++) {
+    for (size_t i = 0; i < board.getX(); i++) {
+        for (size_t j = 0; j < board.getY(); j++) {
             guess(i, j, board);
         }
     }
